`@remotion/studio`: Add asset delete action
